Refactor API into modular subpackages
